Reorganise common test base
